Refactored preview 5 and moved images in folder
